+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - `--max-threads` flag to control the maximum number of threads used for test execution
 
+### Cast
+
+#### Added
+
+- `sncast ledger` command for simple Ledger operations.
+
 ## [0.57.0] - 2026-02-24
 
 ### Forge